% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fern.R
\name{fern_data}
\alias{fern_data}
\title{Download a Ferndale dataset as CSV}
\usage{
fern_data(key, layer = NULL)
}
\arguments{
\item{key}{Dataset key from fern_datasets(), or a raw item_id}

\item{layer}{Layer number (only needed if using raw item_id)}
}
\value{
tibble with dataset records
}
\description{
Download a Ferndale dataset as CSV
}
