% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fern.R
\name{fern_drug_incidents}
\alias{fern_drug_incidents}
\title{Get Ferndale drug incident and NARCAN data}
\usage{
fern_drug_incidents()
}
\value{
tibble with drug incident records (2007-2017)
}
\description{
Get Ferndale drug incident and NARCAN data
}
