% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fern.R
\name{fern_search}
\alias{fern_search}
\title{Search Ferndale open data portal}
\usage{
fern_search(q = "")
}
\arguments{
\item{q}{Search query}
}
\value{
tibble: id, type, title, description, url
}
\description{
Search Ferndale open data portal
}
